# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.RegisterRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - full_name
    - password
    type: object
  entities.Report:
    properties:
      created_at:
        type: string
      data: {}
      date_from:
        type: string
      date_to:
        type: string
      generated_by:
        type: string
      id:
        type: string
      report_type:
        $ref: '#/definitions/entities.ReportType'
      team_id:
        type: string
    type: object
  entities.ReportType:
    enum:
    - rolling_tasks
    - blockers
    - team_summary
    type: string
    x-enum-varnames:
    - ReportTypeRollingTasks
    - ReportTypeBlockers
    - ReportTypeTeamSummary
  entities.RollingTask:
    properties:
      created_at:
        type: string
      first_mentioned:
        type: string
      id:
        type: string
      last_mentioned:
        type: string
      occurrence_count:
        type: integer
      status:
        $ref: '#/definitions/entities.RollingTaskStatus'
      task_description:
        type: string
      team_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  entities.RollingTaskStatus:
    enum:
    - active
    - resolved
    type: string
    x-enum-varnames:
    - RollingTaskActive
    - RollingTaskResolved
  entities.Standup:
    properties:
      blockers:
        items:
          type: string
        type: array
      completed_tasks:
        items:
          type: string
        type: array
      created_at:
        type: string
      date:
        type: string
      id:
        type: string
      in_progress_tasks:
        items:
          type: string
        type: array
      notes:
        type: string
      planned_tasks:
        items:
          type: string
        type: array
      team_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  entities.Team:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      members:
        items:
          $ref: '#/definitions/entities.User'
        type: array
      name:
        type: string
      updated_at:
        type: string
    type: object
  entities.User:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      role:
        $ref: '#/definitions/entities.UserRole'
      team_id:
        type: string
      updated_at:
        type: string
    type: object
  entities.UserRole:
    enum:
    - admin
    - regular
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleRegular
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  standup.SubmitStandupRequest:
    properties:
      blockers:
        items:
          type: string
        type: array
      completed_tasks:
        items:
          type: string
        type: array
      date:
        type: string
      in_progress_tasks:
        items:
          type: string
        type: array
      notes:
        type: string
      planned_tasks:
        items:
          type: string
        type: array
      team_id:
        type: string
      user_id:
        type: string
    required:
    - date
    - team_id
    - user_id
    type: object
  user.CreateTeamRequest:
    properties:
      created_by:
        type: string
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  user.UpdateUserRequest:
    properties:
      full_name:
        type: string
      role:
        $ref: '#/definitions/entities.UserRole'
    type: object
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a Scrum AI backend server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ScrumAI API
  version: "1.0"
paths:
  /admin/users:
    get:
      description: Retrieve a list of all users (Admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: List all users
      tags:
      - admin
  /admin/users/{id}:
    delete:
      description: Delete a user by ID (Admin only)
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete user
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Update user details (Admin or self)
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - users
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - auth
  /auth/logout:
    post:
      description: Invalidate user session
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Logout user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user with email and password
      parameters:
      - description: Registration info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /me:
    get:
      description: Retrieve details of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.User'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get current user info
      tags:
      - users
  /reports/rolling-tasks:
    get:
      description: Retrieve tasks that have been in progress for multiple standups
      parameters:
      - description: Team ID (UUID)
        in: query
        name: team_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.RollingTask'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get team rolling tasks
      tags:
      - reports
  /reports/team-summary:
    get:
      description: Retrieve AI generated team summary reports
      parameters:
      - description: Team ID (UUID)
        in: query
        name: team_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Report'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get team summary reports
      tags:
      - reports
  /standups:
    get:
      description: Retrieve standups for a specific team within a date range
      parameters:
      - description: Team ID (UUID)
        in: query
        name: team_id
        required: true
        type: string
      - description: Start date (RFC3339)
        in: query
        name: date_from
        type: string
      - description: End date (RFC3339)
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Standup'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get team standups
      tags:
      - standups
    post:
      consumes:
      - application/json
      description: Create a new standup entry for the user
      parameters:
      - description: Standup details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/standup.SubmitStandupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Standup'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Submit a daily standup
      tags:
      - standups
  /teams:
    get:
      description: Retrieve a list of all teams
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Team'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: List all teams
      tags:
      - teams
    post:
      consumes:
      - application/json
      description: Create a new team with a name
      parameters:
      - description: Team name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.CreateTeamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Team'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new team
      tags:
      - teams
  /teams/{id}:
    get:
      description: Retrieve details of a specific team by ID
      parameters:
      - description: Team ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Team'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get team details
      tags:
      - teams
  /teams/{id}/members:
    post:
      consumes:
      - application/json
      description: Add a user to a specific team
      parameters:
      - description: Team ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: User ID to add
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add member to team
      tags:
      - teams
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"